Fuck it. Life's too short for dividing this into smaller commits

+++ app/src/main/cpp/tracker/RCF_processor.hpp
@@ -7,7 +7,7 @@
 #include <array>
 #include <cmath>
 
-#include "Balanced_ACF.hpp"
+#include "tracker/fft/Balanced_ACF.hpp"
 #include "math_utilities.hpp"
 
 namespace reBass
@@ -23,12 +23,12 @@ namespace reBass
 
         static constexpr std::size_t OutputLength = WindowLength / 4;
 
-        RCF_processor(double ray_param)
+        RCF_processor(float ray_param)
         {
             for (int n = 0; n < weighting_vector.size(); n++) {
                 weighting_vector[n] =
                         n / (ray_param * ray_param)
-                        * std::exp(
+                        * (float) std::exp(
                                 -1. * n * n
                                 / (2. * std::pow(ray_param, 2.0))
                         );
@@ -36,7 +36,7 @@ namespace reBass
         };
 
         template <typename InputIterator>
-        std::array<double, OutputLength>
+        std::array<float, OutputLength>
         get_rcf(
                 InputIterator input
         ) {
@@ -48,7 +48,7 @@ namespace reBass
             math_utilities::adaptive_threshold(df_frame);
             acf.perform_ACF(df_frame);
 
-            std::array<double, OutputLength> out;
+            std::array<float, OutputLength> out;
             math_utilities::comb_filter(
                     df_frame,
                     weighting_vector,
@@ -62,7 +62,7 @@ namespace reBass
     private:
         std::array<float, WindowLength> df_frame;
         Balanced_ACF<WindowLength> acf;
-        std::array<double, OutputLength> weighting_vector;
+        std::array<float, OutputLength> weighting_vector;
     };
 }
